@@ -12,8 +12,8 @@
 import 'package:serverpod_client/serverpod_client.dart' as _i1;
 import 'greeting.dart' as _i2;
 import 'recipes/recipe.dart' as _i3;
-import 'package:magic_recipe_client/src/protocol/recipes/recipe.dart' as _i4;
-import 'package:serverpod_auth_client/serverpod_auth_client.dart' as _i5;
+import 'package:serverpod_auth_client/serverpod_auth_client.dart' as _i4;
+import 'package:magic_recipe_client/src/protocol/recipes/recipe.dart' as _i5;
 export 'greeting.dart';
 export 'recipes/recipe.dart';
 export 'client.dart';
@@ -43,12 +43,16 @@ class Protocol extends _i1.SerializationManager {
     if (t == _i1.getType<_i3.Recipe?>()) {
       return (data != null ? _i3.Recipe.fromJson(data) : null) as T;
     }
-    if (t == List<_i4.Recipe>) {
-      return (data as List).map((e) => deserialize<_i4.Recipe>(e)).toList()
+    if (t == List<_i4.UserInfo>) {
+      return (data as List).map((e) => deserialize<_i4.UserInfo>(e)).toList()
+          as T;
+    }
+    if (t == List<_i5.Recipe>) {
+      return (data as List).map((e) => deserialize<_i5.Recipe>(e)).toList()
           as T;
     }
     try {
-      return _i5.Protocol().deserialize<T>(data, t);
+      return _i4.Protocol().deserialize<T>(data, t);
     } on _i1.DeserializationTypeNotFoundException catch (_) {}
     return super.deserialize<T>(data, t);
   }
@@ -63,7 +67,7 @@ class Protocol extends _i1.SerializationManager {
     if (data is _i3.Recipe) {
       return 'Recipe';
     }
-    className = _i5.Protocol().getClassNameForObject(data);
+    className = _i4.Protocol().getClassNameForObject(data);
     if (className != null) {
       return 'serverpod_auth.$className';
     }
@@ -84,7 +88,7 @@ class Protocol extends _i1.SerializationManager {
     }
     if (dataClassName.startsWith('serverpod_auth.')) {
       data['className'] = dataClassName.substring(15);
-      return _i5.Protocol().deserializeByClassName(data);
+      return _i4.Protocol().deserializeByClassName(data);
     }
     return super.deserializeByClassName(data);
   }
